feat: Add new configurations and update GitHub Actions workflow
This commit includes the addition of new configuration files for GitlabLint, active-tab-highlighter-v2, and git_toolbox_blame in the .idea directory. Also, the GitHub Actions workflow for derskythe-fw-build has been significantly updated including changes in the env variables, step names, as well as the addition, removal, and rearrangement of several steps within the workflow process.